Create distribute internal document feature
