+++ src/app/GraphQL/Mutations/DistributeDocumentMutator.php
@@ -2,14 +2,11 @@
 
 namespace App\GraphQL\Mutations;
 
-use App\Enums\FcmNotificationActionTypeEnum;
 use App\Exceptions\CustomException;
-use App\Http\Traits\SendNotificationTrait;
+use App\Http\Traits\DistributeToInboxReceiverTrait;
 use App\Models\DocumentSignature;
 use App\Models\Inbox;
 use App\Models\InboxFile;
-use App\Models\InboxReceiver;
-use App\Models\People;
 use App\Models\TableSetting;
 use Carbon\Carbon;
 use Illuminate\Http\Response;
@@ -18,7 +15,7 @@ use Illuminate\Support\Facades\Http;
 
 class DistributeDocumentMutator
 {
-    use SendNotificationTrait;
+    use DistributeToInboxReceiverTrait;
 
      /**
      * @param $rootValue
@@ -90,36 +87,6 @@ class DistributeDocumentMutator
     }
 
     /**
-     * createInboxReceiver
-     *
-     * @param  mixed $inboxId
-     * @param  mixed $args
-     * @return void
-     */
-    protected function createInboxReceiver($tableKey, $inboxId, $stringReceiversIds)
-    {
-        $receivers  = People::whereIn('PeopleId', explode(', ', $stringReceiversIds))->get();
-        foreach ($receivers as $value) {
-            $InboxReceiver = new InboxReceiver();
-            $InboxReceiver->NId           = $inboxId;
-            $InboxReceiver->NKey          = $tableKey;
-            $InboxReceiver->GIR_Id        = auth()->user()->PeopleId . Carbon::now();
-            $InboxReceiver->From_Id       = auth()->user()->PeopleId;
-            $InboxReceiver->RoleId_From   = auth()->user()->PrimaryRoleId;
-            $InboxReceiver->To_Id         = $value->PeopleId;
-            $InboxReceiver->RoleId_To     = $value->PrimaryRoleId;
-            $InboxReceiver->ReceiverAs    = 'to';
-            $InboxReceiver->StatusReceive = 'unread';
-            $InboxReceiver->ReceiveDate   = Carbon::now();
-            $InboxReceiver->To_Id_Desc    = $value->role->RoleDesc;
-            $InboxReceiver->Status        = '0';
-            $InboxReceiver->save();
-        }
-
-        return true;
-    }
-
-    /**
      * createInboxFile
      *
      * @param  mixed $tableKey
@@ -161,27 +128,5 @@ class DistributeDocumentMutator
         ]);
 
         return $response;
-    }
-
-    protected function doSendNotification($inboxId, $args, $stringReceiversIds)
-    {
-        $dept = auth()->user()->role->rolecode_sort;
-        $title = Arr::get($args, 'input.title');
-        $body = $dept . ' telah mengirimkan surat terkait dengan ' . $title . 'Klik di sini untuk membaca dan menindaklanjut pesan.';
-
-        $messageAttribute = [
-            'notification' => [
-                'title' => $title,
-                'body' => str_replace('&nbsp;', ' ', strip_tags($body))
-            ],
-            'data' => [
-                'inboxId' => $inboxId,
-                'groupId' => $inboxId,
-                'peopleIds' => explode(', ', $stringReceiversIds),
-                'action' => FcmNotificationActionTypeEnum::INBOX_DETAIL(),
-            ]
-        ];
-
-        $this->setupInboxReceiverNotification($messageAttribute);
     }
 }
