+++ src/graphql/inbox.graphql
@@ -26,6 +26,11 @@ type UnreadCount {
     draft: Int!
 }
 
+input DistributeInboxInput {
+    inboxId: String!
+    receiversIds: String!
+}
+
 #import people.graphql
 #import documentType.graphql
 #import documentUrgency.graphql
